Move everything to a temp folder for now
