+++ OLD/src/game/drawing/Sprite.java
